Fix item selection on drop
When an item is dropped, FolderView::SelectAndPositionItems() will call
ShellBrowser::SelectItems(). However, that call can occur before the
target items have been added. In that case, the SelectItems() will fail
to select any items.

To fix that, a list of items to select is now maintained. Once an item
from that list is added, it will be selected.