Paste images using the "PNG" clipboard format, when available
This format is used by applications like Chrome, when copying an image.
The advantage over CF_DIB is that transparency is correctly handled,
since the data object will simply refer to the raw PNG data. That also
makes the data simple to handle, since it can be saved directly to disk.