Allow bookmarks in main menu to be right-clicked
This matches the functionality available on the bookmarks toolbar (where
bookmarks shown in the dropdown menu can be right-clicked).

Also, BookmarksMainMenu::OnMenuItemClicked() has been updated to use
BookmarkMenuController::OnMenuItemSelected(). That means that the ctrl
and shift keys can be used to control whether the bookmark is opened in
a new tab and whether that tab is focused.