Add search tabs dialog
This dialog allows the user to search through the set of currently open
tabs. A tab can be searched for both by name and by path. Once an item
is double-clicked, or selected and OK is pressed, the associated tab
will be selected.

This should make it easier to navigate between tabs, especially if there
are a lot of tabs open.

See #325.