Use DarkModeThemeManager in ThirdPartyCreditsDialog
Additionally, version 4.1 of the Rich Edit control is now used. While
this version isn't detected by the resource editor in Visual Studio (it
shows up as a custom control), using this version means that the control
uses the updated text selection color (rather than the older, darker
blue color).

Using the updated version is probably a good idea anyway, particularly
since it was introduced in Windows XP, so there shouldn't be any issues
with it being unavailable.