Use FolderSettings::groupMode in group comparison function
In 408b094a0d5baef09540f5221c37c566860a0072, the application was updated
to allow the sort mode and group mode to be independently set. However,
the group comparison function was still using the sort mode to determine
which type of group comparison to use. That's now been fixed so that the
group mode is used instead.