Owner-draw the main window menu bar
This allows the menu bar to be drawn with dark colors in dark mode. Note
that because disabling the owner-draw style on menu bar items appears to
cause issues (the resulting items aren't correctly spaced), the menu bar
will always be owner-drawn (on Windows 10 and 11), regardless of which
theme is used.

This is part of the work for issue #115.