Set min and max sizes for treeview width
Previously, there were no restrictions on the width of the treeview.
This meant that it could be dragged to the left until the width was less
than 0 or to the right until the width was greater than the parent
window. Neither of those situations is reasonable.

Now, the treeview has both a minimum, absolute, width (in pixels) and a
maximum, percentage, width. Once the width reaches the minimum or the
maximum, it will be capped at that value.

This also means that in cases where the treeview is resized below 0, the
width will be reset to the minimum width. Previously, this resulted in
situations where if the treeview was resized below 0 and the settings
saved, it wasn't possible to show the treeview again from within the
application (see
https://explorerplusplus.com/forum/viewtopic.php?f=2&t=790).

It also means that as the top-level window is sized down, the treeview
may reduce its size as well, to ensure the maximum percentage width
isn't exceeded. That matches the behaviour of Explorer, which will also
reduce the size of its tree during a resize if necessary.

Finally, this change also fixes a small amount of jitter that could
occur when beginning the resize. Previously, the width of the treeview
was determined from the position of the cursor during a drag. However,
since the drag can be started from a range (and not a specific point),
depending on where the drag was started, the edge of the window could
jump as the cursor started to move (to line up with the position of the
cursor).

Now, the distance between the cursor and the edge of the treeview at the
start of the drag is taken into account. When resizing, that distance
will be kept the same.