Fix crash that could occur when closing the application
In de5e931fb2b53c65a24bb628a44477169d6c5282, the application was updated
to call SetFocus() in WM_ACTIVATE. When the focus changes,
MainToolbar::UpdateToolbarButtonStates() will call
Explorerplusplus::CanPaste(), which will indirectly call
IDropTarget::DragEnter() if the clipboard is non-empty. That method can
then result in a call to PeekMessage() being made, which is problematic,
as it will dispatch non-queued messages, in particular WM_CLOSE.

This meant that, previously, if a file had been copied to the clipboard
and the application was closed using the taskbar context menu, it would
crash. That's because it would receive a WM_ACTIVATE message followed by
a WM_CLOSE message. As described above, the handler for WM_ACTIVATE
would ultimately result in a call to PeekMessage() being made. That
would then result in the WM_CLOSE message being dispatched, which would
result in the main window being destroyed and the Explorerplusplus
instance being deleted, all within the call to IDropTarget::DragEnter().
Once that method returned, the application would be in an invalid state
and would crash.

To fix that, the WM_CLOSE handler now simply posts a custom message.
Once that custom message is processed, the application will be closed.
That works because PeekMessage() won't dispatch posted (i.e. queued)
messages. The only way the message can be processed is through the
normal message loop.