Keep track of queued rename item in DirectoryState
Previously, this was tracked within the ShellBrowser class directly and
the queued item was only being cleared if the item to be renamed was
added, or another rename was queued.

That's not correct, as the queued item also needs to be cleared whenever
the directory changes. Without that, the item could be placed briefly
into rename mode if the directory was manually refreshed before the
directory change notification was processed.

To fix that issue, the queued rename item is now tracked in
DirectoryState. That holds items that are only relevant within the
current directory and will be reset whenever the directory changes.