Allow main font to be customised
The font can be chosen from a new "Fonts" page in the options dialog.
Currently, the font size is capped between 6pt and 24pt. Allowing
smaller sizes isn't very useful. Allowing larger sizes might be, but
trying to use a very large font size isn't going to work well in
practice. It means that there won't be enough room on screen for the UI
elements. So, allowing arbitrarily large font sizes likely isn't
reasonable.

If set, this font will be used by:

- The address bar
- Each toolbar
- Each listview
- The treeview
- The tooltip windows for each of those controls

Because choosing a custom font won't result in any other changes being
made, choosing a font that's significantly different from the standard
font can result in a UI that looks mismatched. For example, icons will
still be drawn at their regular size, no matter how large the font size
of the custom font. So, a custom font works better when it's closer to
the standard font.

This is part of the work for #27.