Update font used to measure main menu text
The font retrieved via GetThemeSysFont() is typically correct, however,
it won't be right when the text size setting is changed in Windows. In
that situation, the font returned by GetThemeSysFont() will still be the
original size.

Therefore, SystemParametersInfoForDpi() will now be used to retrieve the
menu font.

This fixes an issue where the menu bar items wouldn't be correctly
spaced if the text size was increased.