Push build artifacts to S3
AppVeyor only keeps build artifacts for a month, which is problematic.

For example, there have been cases where someone has reported a crash
and provided a crash dump, but the build artifacts have already expired
(so the debugging symbols are no longer available).

There have also been multiple cases where the artifacts have expired and
the download links for the dev build have broken.

Deploying the artifacts to S3 will allow them to be kept indefinitely.

Doing that also makes it easier for users to switch to an earlier dev
build in the case that a bug is introduced.

At the moment, only dev builds are deployed.