Perform exact version match when loading translation DLL
Previously, only the major, minor and micro version were checked.
However, the only version number that has actually changed for a while
now is the build version. That meant that not checking the build version
allowed older translation DLLs to be loaded, which could cause crashes
if a specific resource wasn't available (see #396).

Therefore, an exact version comparison is now performed when loading a
translation DLL. While doing that should prevent the problem in almost
every situation, it's still possible for a translation DLL to be out of
date, even if it's for the exact same version.

When pulling updated translations in from Crowdin, the repository needs
to be updated first. That means that if a resource is added, for
example, the repository will be updated, AppVeyor will build the code
and the updated translations will be pulled from Crowdin. The repository
will then be updated again and another build will be launched on
AppVeyor. In that situation, only the second build will have up-to-date
translation DLLs. So, it would still be worthwhile to put other
mitigations in place.