Correct errorlevel set by the "set" command
As per the help page for the "set" command, errorlevel will be set to 1
if the variable name isn't in the current environment. That behaviour
then causes AppVeyor to fail, since it treats an exit code of 1 as an
error.

To fix that, a dummy command is now run after the "set" command. That
dummy command will reset the errorlevel.