Mark double click as handled when navigating up
In f1ddd6cc8302914c7d3acf1f4a9008058db0ca58, the LBUTTONDBLCLK handler
was moved to the ShellBrowser class. Originally, the handler returned 0
when navigating up, to indicate that the double click message had been
handled. After the code had been moved, however, DefSubclassProc was
always invoked, which meant that NM_DBLCLK would also be sent.

Overall, this meant that if you double clicked in an empty area, an up
navigation would occur (due to the LBUTTONDBLCLK handler) and,
confusingly, the item under the mouse would then be opened (by the
NM_DBLCLK handler).

To fix that, the LBUTTONDBLCLK handler has been updated to return 0 when
navigating up.

This fixes #390.