Update download links for latest builds in README.md
The download links have been updated to point to the artifacts stored in
S3 (served from download.explorerplusplus.com). At the moment, only the
latest version is linked to in the readme file. In future, a page will
be added to explorerplusplus.com that will list all files available in
the bucket. At that point, a link to the page will be added to the
readme.

Also, the symbol links have been removed, since there's not much of a
need for them to be in the readme.