Ensure build number is set for translation DLLs
In 45af8875b22f39c7c93353af7735c5fc57327415, the code was updated to
perform an exact version check for translation DLLs. However, the DLLs
were only being stamped with the major, minor and micro versions; the
build version was always set to 0. That then meant that it wasn't
possible to load a translation DLL for an AppVeyor build, since the
version numbers wouldn't match.

To fix that, the APPVEYOR_BUILD_NUMBER environment variable (if
available) is now used to set the build version, in the same way the
variable is used in the Explorer++ project.