Correct group box text spacing in dark mode
Previously, there was some hard-coded spacing between the box outline
and the text. However, that spacing was only applied to the left side of
the text. Now, the value is retrieved dynamically and the spacing is
applied to the left and right of the text.

This is part of the work for issue #115.