Deploy latest_version.txt only once at end of build process
The latest_version.txt artifact is now only deployed a single time, at
the end of the ARM64 build. Previously, the file was deployed at the end
of each platform build (Win32, x64 and ARM64). As the link to download
the latest version depends on latest_version.txt, that behaviour then
meant that links for the latest version would temporarily break during
the build process (while the build for a particular platform was still
in progress).

Updating the file a single time at the end of the build process ensures
that the links are only updated once each platform build is finished.

The method used here isn't ideal, since it implicitly depends on the
fact that the ARM64 build is the last build, but it should work ok.