Update NM_CLICK/NM_DBLCLK handler to use correct mouse coordinates
Previously, the handler used for NM_CLICK/NM_DBLCLK called
GetMessagePos() to retrieve the mouse coordinates. While that typically
worked, it didn't work in the situation where one click activate was on
and the mouse was dragged before being released.

In that situation, the NMITEMACTIVATE struct passed to NM_CLICK would
contain the coordinates of the mouse at the start of the drag. However,
GetMessagePos() would return the coordinates at the end of the drag.
Since those coordinates would be used to determine which item to open,
that then meant that the item under the mouse at the end of the drag
would be opened.

To fix that, the coordinates specified in the NMITEMACTIVATE struct are
now used directly. That also simplifies the handler.

This fixes #375.