Select root item in treeview by default
If the "Automatically synchronize treeview with main pane" option is
disabled, the treeview selection won't be updated when a navigation
occurs.

Previously, that also meant that the treeview would have no initial
selection. Then, when the treeview gained focus, it would automatically
select an item, which would cause a navigation to that item to occur.

To fix that behaviour, the root item is now selected by default. That
ensures that if the focus is shifted to the treeview, the selection
won't unexpectedly change.