Allow main font size to be decreased/increased directly
The font size can now be decreased/increased via the shortcut keys
Ctrl + - and Ctrl + +. When one of those shortcut keys is pressed, the
main font will be decreased/increased by 2 points at a time. Two menu
entries to adjust the font size have also been added to the View menu.

If no custom font is currently active, a custom font will be set up
based on the system font, with the size adjusted as necessary.

This is part of the work for #27.