Reinsert toolbar buttons when DPI is changed
When the DPI is changed, the toolbar control won't update the button
sizes (which may lead to text being cut off if the DPI is increased).
Additionally, the height of the toolbar won't change, which can cause
the buttons not to be shown at all if the height isn't sufficient.

To fix this, the toolbar buttons will now be removed and reinserted when
the DPI changes. This matches the solution used when the main font is
changed.