Remove option to use fixed width tabs
When fixed width tabs are turned on, the tab control doesn't always
redraw tabs correctly. For example, the icon and text can overlap (or
the icon can potentially hide the text if it's short). Additionally, the
alignment can be potentially wrong - a tab can be center-aligned instead
of left-aligned and vice versa. That can be seen when dragging a tab
into a different position. These issues are present from at least
Windows XP - Windows 11.

Since it appears the control simply doesn't properly support this
option, it doesn't make much sense to expose it to users. While it is
possible to work around the issues the style causes, by forcing the tab
control to redraw using a WM_SIZE message, that's not a particularly
good solution. It adds additional complexity and it might be difficult
to ensure that WM_SIZE is always called when necessary.

Therefore, the option to use fixed width tabs has been removed.

This fixes #255.