Start implementation of dual pane functionality
A "Dual pane" item has been added to the View menu. That menu entry will
be removed if the dual panes functionality is disabled, which is
currently the case - the functionality is disabled at compile time.
Therefore, there's no user-facing change here at the moment.

This is part of the work for #291.