Update mousewheel handling
Previously, there was some code to forward mousewheel messages sent to
the listview, treeview and address bar. If the window under the mouse
wasn't the target window, the message would be forwarded to the window
at the mouse position (provided it was the listview or treeview).

This was been used on versions of Windows prior to Windows 10 in order
to try and scroll the window under the mouse, rather than the window
with focus, if the two were different.

That handling has now been removed. That simplifies mousewheel handling
and makes the application consistent with the underlying OS behaviour.

Note that this does mean that there's a behaviour change on anything
prior to Windows 10 - scrolling the mousewheel will now always scroll
the window that has focus. On Windows 10 and above, the behaviour should
be unchanged.