Ensure view mode is selected on background context menu in Windows 7
Previously, when showing the background context menu on Windows 7, the
view modes were being manually added to the menu. However, the current
view mode wasn't being checked.

Now, the views submenu is created and attached to the context menu in
the same way as its created and attached when showing the background
context menu on Windows 8 and above. This ensures the current view mode
is checked.

Note that, with this change, the "Auto Arrange" item is no longer shown
on the views submenu on Windows 7. That's consistent with how the menu
is shown on Windows 8 and above.