Add .readthedocs.yaml config file
This config file is now required to build the documentation on Read the
Docs (see https://blog.readthedocs.com/migrate-configuration-v2/).
Recent documentation builds have started failing due to the absence of
the file.

As part of this change, the documentation has also been updated to build
with the latest version of Sphinx. As recommended by Read the Docs, the
dependencies have also been pinned, so that specific versions will
always be used for the build.