Decrease/increase main font size by 1 point at a time
Previously, when using Ctrl + - and Ctrl + +, the font size would be
decreased/increased by 2 points at a time. However, that meant that if
the font reached the minimum/maximum, it wasn't possible to get back to
the default font size (9 point), since the font size would always be
even. Changing the size in 1 point increments ensures that the default
font size can always be reached.