Show home folder in treeview on Windows 11
The home folder was added in Windows 11 and has very similar contents to
the quick access folder that was added in Windows 10. Previously, the
quick access folder was always shown at the top of the treeview.
However, on Windows 11, the home folder will now be shown instead. That
makes more sense on Windows 11, since:

1. That's the folder shown in Explorer.
2. When an item is pinned/unpinned, the change notification is generated
   for the home folder (rather than the quick access folder). Although
   the treeview doesn't currently handle that notification type
   (SHCNE_UPDATEDIR), showing the home folder means it's at least
   possible to update the treeview when necessary.

On Windows 10, the quick access folder will still be shown.