Handle SHCNE_UPDATEDIR in treeview for quick access folder specifically
The SHCNE_UPDATEDIR notification indicates that the contents of a folder
have changed. It's difficult to handle in the general case, since
there's no specific information about what changed. However, handling
the notification specifically for the quick access folder is possible,
since there will only be a small number of items at a single level.

Therefore, when the SHCNE_UPDATEDIR notification is received for the
quick access folder, the node in the treeview will be collapsed and
re-expanded. That now means that pinning or unpinning an item will
result in the quick access folder in the treeview being updated.