When navigating to a shell link, navigate to target folder instead
Various items in the shell namespace are links (i.e. they support the
IShellLink interface). Examples of these items include:

- Shortcuts
- Symlinks
- Virtual link objects (e.g. pinned items in the quick access folder)

When navigating to a shell link, it can be important to navigate to the
target instead. An example of why that's important is the quick access
folder. That folder contains recent items as well as items that the user
has pinned. Although navigating directly to a recent/pinned folder works
as expected, directory monitoring doesn't work. Presumably, that's
because directory change notifications are only generated for the
original directory. To have things work correctly, the navigation needs
to proceed to the target folder instead.

One thing to note is that Explorer itself is inconsistent in its
behaviour. For example, clicking a symlink (e.g. C:\Users\All Users) in
the navigation pane will cause a navigation to the target (e.g.
C:\ProgramData). However, clicking a symlink (e.g. C:\Users\All Users)
in the main pane will result in a navigation to the symlink (e.g.
C:\Users\All Users). The contents displayed are the same, since the
symlink and target are effectively two names for the same thing, but
it's still inconsistent.

Here, the decision has been made to always navigate to the target, if
one is available (i.e. if the object supports IShellLink). That doesn't
change how shortcuts are handled, since they're currently not passed
directly to ShellBrowser::Navigate(). It does, however, change how
symlinks and virtual links are handled. For both of those cases, a
navigation to the link will now result in a navigation to the target
instead.

A practical benefit of doing that is that it means that directory change
notifications now work when opening a recent/pinned folder under quick
access.

This fixes #405.