Fix deletion of treeview items
When deleting a treeview item via the main menu or the delete key, the
deletion would silently fail. That's likely because
ExecuteActionFromContextMenu() was changed in
9921834cb0244bc4ebac0400fcdc9f0cb88738ba to retrieve the background
context menu for a folder when not passing any items (rather than the
context menu for the folder). Because the background context menu
doesn't have a delete item, attempting to invoke the delete item would
have no effect.

To fix that, files in the treeview are now deleted in the same way
they're deleted in the listview (i.e. using
FileActionHandler::DeleteFiles()).

This fixes #407.