Simplify code used to remove treeview item
This also fixes an issue where, if the current folder was deleted in the
treeview and the selection was shifted to the parent, no navigation
would occur. That's because, previously, when an item was deleted, if
there were no other items in the parent folder, the parent would simply
be collapsed, without explicitly deleting the removed item. The treeview
would automatically shift the focus to the parent, but without sending a
TVN_SELCHANGED notification. That then meant that the current folder
wouldn't change.

Now, the deleted item is always explicitly removed, which will generate
a TVN_SELCHANGED notification and will cause a navigation to occur.