Handle Ctrl+Backspace in listview edit control
Edit controls don't handle Ctrl+Backspace in the way that's typically
expected (i.e. by erasing the last word). Instead, a non-printable
character is inserted. This is due to the fact that pressing
Ctrl+Backspace generates a WM_CHAR message, with the character code set
to 127 (i.e. the delete non-printing control code).

To fix that, the WM_CHAR message is now ignored when the character code
is 127. Additionally, when Ctrl+Backspace is pressed and there's no text
currently selected, a Ctrl+Shift+Left key press will be sent to the edit
control. This will cause it to select the last word. The selection will
then be erased.

Two alternative solutions were also considered:

- Manually deleting the last word on Ctrl+Backspace, using something
  like a regex. While that's a more direct way of erasing the last word,
  trying to handle unicode is complicated. The simple solution doesn't
  work for anything but ascii characters.
- Using ICU. ICU makes it trivial to break a unicode string up into a
  set of words and erase the last word. However, it also increases the
  executable size from 5MB to 35MB and is overkill, since most of the
  functionality in the library isn't needed.

Therefore, the current solution was chosen. While it's hacky, it does
also have the advantage that Ctrl+Backspace functions in the same way as
Ctrl+Del, since the edit control is always responsible for finding the
previous/next word.

This fixes #299.