Handle undo keypress (Ctrl+Z) outside of accelerator system
Previously, the undo item appeared in the accelerator table. The
downside of that was that it meant that other controls could never
handle the key press, since they would never receive it.

Now, Ctrl+Z is handled by the listview specifically. That means that
other controls (e.g. the address bar, listview edit control) can process
Ctrl+Z when it's pressed.