Update item name in listview when SHCNE_UPDATEITEM is received
Item updates and renames are handled by the same code in the
ShellBrowser class. Previously, the displayed item name was only updated
if a rename notification was being processed. However, that's actually
not correct.

When a virtual item is renamed, it's possible that only the display name
changes, while the parsing name remains the same (e.g. that's the case
for the recycle bin). In that case, the only notification that will be
generated is a SHCNE_UPDATEITEM notification and the displayed name
should be updated in response to it.

Therefore, the displayed name is now always updated when handling
SHCNE_UPDATEITEM.