Remove calls to SetCurrentDirectory()
The rationale for adding these calls is given in
https://explorerplusplus.com/forum/viewtopic.php?t=663. That is, some
context menu items may rely on the current directory being set.
However, it's no longer necessary to do that, as the working directory
is set when calling IContextMenu::InvokeCommand().

Since SetCurrentDirectory() maintains a handle to the directory,
removing these calls also means that it's now possible to delete a
directory, even if Explorer++ has it open. Note that, that situation
isn't fully handled at the moment.

If the treeview is shown and treeview synchronisation is on, the
treeview will select another folder when the current folder is deleted
and a navigation will occur. If the treeview isn't shown or treeview
synchronisation isn't on, the treeview won't select another folder and
no navigation will occur.