Fix crash that could occur when processing SHCNE_UPDATEDIR notification
When processing SHCNE_UPDATEDIR in the ShellBrowser class, the current
directory will be refreshed. However, it's not safe to do that
immediately, as refreshing will clear
ShellChangeWatcher::m_shellChangeNotifications, which is being actively
iterated through.

To fix that, a request to refresh the directory is now posted to the
message queue instead.