Navigate up if current folder is indirectly deleted
If a folder is indirectly deleted, by deleting a parent folder, no
change notification will be sent for the folder itself. That meant that,
previously, if treeview synchronisation was off, no navigation would
occur if the folder was indirectly deleted.

Now, folder deletions are monitored for the entire tree by each tab.
This is the same method used by Explorer to detect deletion of the
current folder. If a parent of the current folder is deleted, a
navigation will occur to the closest existing parent, similar to what's
done in Explorer.