Perform additional validation of IShellFolder::SetNameOf()
This method can return S_OK, even when no rename took place and the
output child pidl is null. That meant that, in the treeview, if you
renamed an item without changing the name, the item would be updated
with the parent's details.

That's because IShellFolder::SetNameOf() would return S_OK, along with a
null child pidl. That child pidl would then be combined with the parent
pidl, which would produce the pidl of the parent. The child would then
be updated with that pidl, which would give it the parent's details.

To fix that, the output child pidl is now checked when calling
IShellFolder::SetNameOf(), along with the result code. This ensures that
if you don't change the item name, nothing is updated.