Allow navigations to override tab lock
This is useful, for example, for navigations that need to take place
after the current folder has been renamed or deleted. In those sorts of
situations, a navigation has to occur in a particular tab, regardless of
whether or not that tab is locked.

This allows several navigations to go through the
ShellNavigationController class, rather than going directly to
ShellBrowser::Navigate().

This change also ensures better consistency when navigating to the
current location in a locked tab. Previously, some navigations to the
same location would open a new tab (e.g. when pressing enter in the
address bar or when clicking a bookmark). Now, all navigations to the
same location in a locked tab proceed in that tab (i.e. the navigation
is treated as an implicit refresh). The only exception is the history
navigations, which will open in a new tab, even if the history entry is
for the same location as the current entry.