Ensure history entry is always added for first navigation
This fixes a crash that would occur when attempting to navigate back in
a locked tab. The previous history entry would be opened in a new tab,
except no history entry would be added, since the navigation was
designated as a history navigation. The application would then crash,
due to the fact that there always needs to be a current history entry.

To fix that, a history entry will now always be added on the first
navigation, regardless of the type requested as part of the navigation.