Always replace existing history entries
When navigating to a history entry, it's important to update that entry.
For example, the display name of a shell item might change, even if the
parsing path doesn't. That means the history item can still be valid,
while its data may be out of date.

Replacing the history entry means that a situation like the following
will work correctly:

1. A navigation to the recycle bin occurs.
2. A navigation to another folder occurs.
3. The name of the recycle bin is updated.
4. The tab is navigated back to the recycle bin. Components that
   retrieve and display entry details (e.g. the address bar) will then
   display the correct, updated details.