Navigate to new directory after parent rename
When a parent is renamed, a SHCNE_UPDATEDIR notification will be
generated. That notification is now used to check whether or not the
current folder still exists. If it doesn't exist, a navigation to the
closest existing parent will occur.

SHCNE_UPDATEDIR may also be sent when a parent folder is deleted, so the
above also applies in that situation.