Show dropdown menu for up toolbar button
This menu will display each folder above the current folder. The items
in the menu can be both clicked (which will open the item in the current
tab) and middle-clicked (which will open the item in a new tab). The
ctrl and shift keys can also be used to control exactly how an item is
opened.