Move OnShowTreeViewContextMenu() to ShellTreeView class
This also fixes an issue in which holding shift while right-clicking a
drive on the drives toolbar, or an item in the search dialog, would show
the rename item, rather than the extended menu. That's because, in both
of those cases, the arguments being passed to
FileContextMenuManager::ShowMenu() were subtly wrong.

e5ca4ba229cdc6765ca11ec258f0f68b7b83f9ad added an additional IUnknown*
parameter to FileContextMenuManager::ShowMenu(). However, the calls to
that method in the drives toolbar and search dialog weren't updated, so
they incorrectly passed in the rename/extended flags. This change
resolves that issue.