Use Google Logging (glog) for logging
Previously, boost-log was being used. Using glog means that the DCHECK
and CHECK macros can be used. CHECK is particularly useful, since it
means that the previous method of reacting to an invalid assumption
(typically by throwing a deliberately unhandled exception) no longer
needs to be used.

Additionally, when logging is enabled, any CHECK failure will be written
to the log file. That means that even if a minidump isn't created when a
CHECK is triggered, the log file will give the exact line which failed.

By default, logging is disabled. Previously, it was enabled by default
in debug builds, but that's not very useful, since there's almost always
better ways of getting information when running a debug build. In this
case, disabled means that the logs are only written to stdout and only
when the severity is at least ERROR. That effectively means that logging
will be disabled from the user's point of view.

When logging is enabled, the default log folder and log filename will be
used. In practice, that means that logs will be written to %TEMP%.
That's generally better, since the application directory may not be
writeable.