Fix crash that could occur when dropping an item in the tab container
When a drag is performed over the tab control, the TabContainer class
sets a timer to change the selected tab. That way, a tab will be
selected if an item is dragged over it, though the selection won't
change immediately.

The issue with that, however, is that WM_TIMER messages can remain in
the queue, even after KillTimer() has been called. The reason for that
is explained in
https://devblogs.microsoft.com/oldnewthing/20141205-00/?p=43463.
Essentially, if PeekMessage() is called with the PM_NOREMOVE flag, a
WM_TIMER message can be generated, but not immediately handled.

When a drag and drop operation is being performed, COM can call
PeekMessage() with the PM_NOREMOVE flag. That can result in a call
sequence like the following:

SetTimer()
When an item is dragged over a tab.

PeekMessage()
This will be called as part of the COM modal loop.

KillTimer()
When the item is dragged over a blank part of the tab control or outside
of the tab control.

WM_TIMER
This message was generated by the call to PeekMessage(), but wasn't
immediately handled.

Previously, that sequence was considered to be invalid by the tab
control. If it occurred, m_dropTargetIndex would be reset to -1 when the
timer was killed. When the WM_TIMER message was then handled, the tab
control would attempt to select the tab at index -1, which would fail
and trigger a crash.

To fix that, a pair of classes OneShotTimerManager and OneShotTimer have
been added. These classes will ensure that a timer is never called after
the timer has been stopped. The TabContainer class has been updated to
use those classes.

This fixes #428.