Move back/forward/up command handlers to new class
This has two specific advantages:

1. It helps to reduce the size of the Explorerplusplus class, which is
   still doing too much.
2. It makes it easier to call the methods from other classes, without
   having to add them to CoreInterface. CoreInterface essentially
   aggregates various unrelated methods and shouldn't be added to.

With this change, the ctrl and shift keys can now also be used to modify
the disposition used when clicking the back/forward/up items on the main
menu or toolbar. Note, though, that the main toolbar currently uses the
shift key to allow buttons to be rearranged. So that means that it's not
possible to use the modifiers shift or ctrl+shift. However, the ctrl
modifier will still work for the main toolbar, as will the shift
modifier when middle-clicking.