Fix issue with previously selected items not being re-selected after navigation
In f90cf5c503607fbf6dd566e0cde014966ce471da, the application was updated
to always replace existing navigation entries. That then meant that when
ShellBrowser::OnEnumerationCompleted() attempted to look up a history
entry to re-select any previously selected items, the lookup would fail.
That's because the history entry would always be replaced before the
call to ShellBrowser::OnEnumerationCompleted().

To fix that, the set of selected items will now be copied when an entry
is replaced. Additionally, ShellBrowser::OnEnumerationCompleted() will
retrieve the set of selected items from the current history entry,
rather than from the entry that the navigation was based on.

That should work ok, since the current entry won't have any selected
items if it represents a new navigation to a folder. There will only be
selected items for a previous history entry.