Only package artifacts for release build on AppVeyor
The Debug-Asan configuration is only enabled on AppVeyor so that the
tests can be run while ASAN is enabled. The artifacts aren't really
useful. Even if an ASAN build might help track down an issue, it
wouldn't run without the ASAN DLL anyway, which is only included with
Visual Studio.

Therefore, artifacts are now only packaged if building in release mode.