Use PowerShell Core on AppVeyor, rather than Windows PowerShell
In bf7fda7044286eb94a4b7d24c48fe9ebcacc0325, the AppVeyor build process
was updated to use PowerShell for the after_build script. As part of
that, Out-File was used to write the build version to
latest_version.txt.

In Windows PowerShell, the default encoding for Out-File is UTF-16 LE
BOM. That then meant that the AWS lambda used to retrieve the latest
version was failing, because it expected the file to be in UTF-8. Which
then broke the links for the latest dev build.

To fix that, PowerShell Core is now used on AppVeyor, rather than
Windows PowerShell. In PowerShell Core, the default encoding for
Out-File is UTF-8.