Refactor application toolbar drag and drop handling code
The code to handle drag and drop operations for the application toolbar
has been moved into a new ApplicationDropper class. Several tests have
been added for this class.

Additionally, this also fixes some unusual behaviour on at least Windows
7, where the toolbar would indicate that a drop would succeed during the
drag operation, but then the drop would silently fail, due to the fact
that retrieving the PKEY_ItemNameDisplayWithoutExtension property would
fail. In a similar vein, dropping an item with just an extension (e.g. a
file named ".exe") would fail for the same reason.

To fix both of those issues, the application name is now derived from
the display name of the file, with the full name being used if the name
just contains the extension.

Also, with the way the ApplicationDropper class works, it's no longer
possible to have a situation where the drag indicates that it would
succeed, but the drop would then fail. That's because all the
information needed for the drop to succeed is extracted at the start of
the drag operation, rather than when the drop occurs.

Finally, there are two other smaller changes that have been made:

- When dragging items over an existing application, the drag description
  will be changed to "Open with [application path]".
- When Explorerplusplus::OpenFileItem() is called, it will use the path
  of the folder opened in the active ShellBrowser (if non-virtual) as
  the start directory for the file, rather than the path of the file's
  parent.
  That's generally better, as it means that when an application from the
  applications toolbar is opened, for example, the start directory will
  be that of the current folder, which is in line with what would be
  expected.